+++ application/src/main/kotlin/com/hanghe/redis/movie/MovieService.kt
@@ -3,9 +3,8 @@ package com.hanghe.redis.movie
 import com.hanghe.redis.cache.CacheInfo
 import com.hanghe.redis.cache.CacheManager
 import com.hanghe.redis.movie.response.GetMovieScreeningResponses
-import com.hanghe.redis.mysql.movie.MovieRepository
-import com.hanghe.redis.mysql.screening.ScreeningRepository
 import com.hanghe.redis.ratelimiter.RateLimiter
+import com.hanghe.redis.screening.ScreeningRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
