+++ infrastructure/build.gradle.kts
@@ -1,16 +1,16 @@
-dependencies {
-    implementation("org.springframework.boot:spring-boot-starter-data-jpa:3.4.3")
-    implementation("mysql:mysql-connector-java:8.0.33")
-    implementation("org.springframework.boot:spring-boot-starter-data-redis:3.4.3")
-
-    // jackson
-    implementation("com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.15.2")
-    implementation("com.fasterxml.jackson.datatype:jackson-datatype-jdk8:2.15.2")
-
-    // QueryDSL
-    implementation("com.querydsl:querydsl-jpa:5.0.0:jakarta")
-    kapt("com.querydsl:querydsl-apt:5.0.0:jakarta")
-
-    implementation(project(":domain"))
-    implementation(project(":common"))
-}
+//dependencies {
+//    implementation("org.springframework.boot:spring-boot-starter-data-jpa:3.4.3")
+//    implementation("mysql:mysql-connector-java:8.0.33")
+//    implementation("org.springframework.boot:spring-boot-starter-data-redis:3.4.3")
+//
+//    // jackson
+//    implementation("com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.15.2")
+//    implementation("com.fasterxml.jackson.datatype:jackson-datatype-jdk8:2.15.2")
+//
+//    // QueryDSL
+//    implementation("com.querydsl:querydsl-jpa:5.0.0:jakarta")
+//    kapt("com.querydsl:querydsl-apt:5.0.0:jakarta")
+//
+//    implementation(project(":domain"))
+//    implementation(project(":common"))
+//}
