+++ infrastructure/cache/build.gradle.kts
@@ -1,0 +1,7 @@
+dependencies {
+    implementation("org.springframework.boot:spring-boot-starter-data-redis:3.4.3")
+
+    // jackson
+    implementation("com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.15.2")
+    implementation("com.fasterxml.jackson.datatype:jackson-datatype-jdk8:2.15.2")
+}
