+++ infrastructure/cache/src/main/kotlin/com/hanghe/redis/cache/CachePair.kt
