+++ infrastructure/message/build.gradle.kts
@@ -1,0 +1,5 @@
+dependencies {
+    implementation("org.springframework.boot:spring-boot-starter-data-jpa:3.4.3")
+
+    implementation(project(":domain"))
+}
