+++ infrastructure/mysql/build.gradle.kts
@@ -1,0 +1,10 @@
+dependencies {
+    implementation("org.springframework.boot:spring-boot-starter-data-jpa:3.4.3")
+    implementation("mysql:mysql-connector-java:8.0.33")
+
+    // QueryDSL
+    implementation("com.querydsl:querydsl-jpa:5.0.0:jakarta")
+    kapt("com.querydsl:querydsl-apt:5.0.0:jakarta")
+
+    implementation(project(":domain"))
+}
