+++ infrastructure/mysql/src/main/kotlin/com/hanghe/redis/movie/MovieJpaRepository.kt
@@ -1,0 +1,5 @@
+package com.hanghe.redis.movie
+
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface MovieJpaRepository : JpaRepository<MovieEntity, Long>
