+++ infrastructure/mysql/src/main/kotlin/com/hanghe/redis/movie/MovieRepository.kt
@@ -1,0 +1,8 @@
+package com.hanghe.redis.movie
+
+interface MovieRepository {
+
+    fun findAll(): List<MovieEntity>
+
+    fun search(title: String?, genre: String?): List<MovieEntity>
+}
