+++ infrastructure/mysql/src/main/kotlin/com/hanghe/redis/screening/ScreeningRepository.kt
@@ -1,0 +1,12 @@
+package com.hanghe.redis.screening
+
+interface ScreeningRepository {
+
+    fun findByMovieIdOrderByStartTime(movieId: Long): List<ScreeningEntity>
+
+    fun saveAll(movies: List<ScreeningEntity>): List<ScreeningEntity>
+
+    fun findById(screeningId: Long): ScreeningEntity?
+
+    fun getById(screeningId: Long): ScreeningEntity
+}
