+++ infrastructure/mysql/src/main/kotlin/com/hanghe/redis/seat/SeatJpaRepository.kt
@@ -1,0 +1,9 @@
+package com.hanghe.redis.seat
+
+import com.hanghe.redis.movie.seat.SeatEntity
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface SeatJpaRepository : JpaRepository<SeatEntity, Long> {
+
+    fun findByTheaterIdAndIdIn(theaterId: Long, seatIds: List<Long>): List<SeatEntity>
+}
