+++ infrastructure/mysql/src/main/kotlin/com/hanghe/redis/theater/TheaterJpaRepository.kt
@@ -1,0 +1,6 @@
+package com.hanghe.redis.theater
+
+import com.hanghe.redis.theater.TheaterEntity
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface TheaterJpaRepository : JpaRepository<TheaterEntity, Long>
