+++ infrastructure/mysql/src/main/kotlin/com/hanghe/redis/theater/TheaterRepository.kt
@@ -1,0 +1,6 @@
+package com.hanghe.redis.theater
+
+interface TheaterRepository {
+
+    fun findAll(): List<TheaterEntity>
+}
