+++ infrastructure/mysql/src/main/kotlin/com/hanghe/redis/theater/TheaterRepositoryImpl.kt
@@ -1,0 +1,13 @@
+package com.hanghe.redis.theater
+
+import org.springframework.stereotype.Repository
+
+@Repository
+class TheaterRepositoryImpl(
+    private val theaterJpaRepository: TheaterJpaRepository
+) : TheaterRepository {
+
+    override fun findAll(): List<TheaterEntity> {
+        return theaterJpaRepository.findAll()
+    }
+}
