--- infrastructure/src/main/kotlin/com/hanghe/redis/mysql/movie/MovieJpaRepository.kt
@@ -1,6 +1,0 @@
-package com.hanghe.redis.mysql.movie
-
-import com.hanghe.redis.movie.MovieEntity
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface MovieJpaRepository : JpaRepository<MovieEntity, Long>
