--- infrastructure/src/main/kotlin/com/hanghe/redis/mysql/movie/MovieRepository.kt
@@ -1,10 +1,0 @@
-package com.hanghe.redis.mysql.movie
-
-import com.hanghe.redis.movie.MovieEntity
-
-interface MovieRepository {
-
-    fun findAll(): List<MovieEntity>
-
-    fun search(title: String?, genre: String?): List<MovieEntity>
-}
