--- infrastructure/src/main/kotlin/com/hanghe/redis/mysql/reservation/ReservationRepositoryImpl.kt
@@ -1,33 +1,0 @@
-package com.hanghe.redis.mysql.reservation
-
-import com.hanghe.redis.reservation.ReservationEntity
-import org.springframework.stereotype.Repository
-
-@Repository
-class ReservationRepositoryImpl(
-    private val repository: ReservationJpaRepository,
-) : ReservationRepository {
-
-    override fun findByScreeningId(screeningId: Long): List<ReservationEntity> {
-        return repository.findByScreeningId(screeningId)
-    }
-
-    override fun saveAll(newReservations: List<ReservationEntity>) {
-        repository.saveAll(newReservations)
-    }
-
-    override fun findByScreeningIdAndSeatIds(screeningId: Long, seatIds: List<Long>): List<ReservationEntity> {
-        return repository.findByScreeningIdAndSeatIdIn(screeningId, seatIds)
-    }
-
-    override fun findByScreeningIdAndSeatIdsWithPessimisticLock(screeningId: Long, seatIds: List<Long>): List<ReservationEntity> {
-        return repository.findByScreeningIdAndSeatIdsWithPessimisticLock(screeningId, seatIds)
-    }
-
-    override fun findByScreeningIdAndSeatIdsWithOptimisticLock(
-        screeningId: Long,
-        seatIds: List<Long>
-    ): List<ReservationEntity> {
-        return repository.findByScreeningIdAndSeatIdsWithOptimisticLock(screeningId, seatIds)
-    }
-}
