--- infrastructure/src/main/kotlin/com/hanghe/redis/mysql/screening/ScreeningRepository.kt
@@ -1,14 +1,0 @@
-package com.hanghe.redis.mysql.screening
-
-import com.hanghe.redis.screening.ScreeningEntity
-
-interface ScreeningRepository {
-
-    fun findByMovieIdOrderByStartTime(movieId: Long): List<ScreeningEntity>
-
-    fun saveAll(movies: List<ScreeningEntity>): List<ScreeningEntity>
-
-    fun findById(screeningId: Long): ScreeningEntity?
-
-    fun getById(screeningId: Long): ScreeningEntity
-}
