--- infrastructure/src/main/kotlin/com/hanghe/redis/mysql/seat/SeatJpaRepository.kt
@@ -1,9 +1,0 @@
-package com.hanghe.redis.mysql.seat
-
-import com.hanghe.redis.movie.seat.SeatEntity
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface SeatJpaRepository : JpaRepository<SeatEntity, Long> {
-
-    fun findByTheaterIdAndIdIn(theaterId: Long, seatIds: List<Long>): List<SeatEntity>
-}
