--- infrastructure/src/main/kotlin/com/hanghe/redis/mysql/seat/SeatRepository.kt
@@ -1,12 +1,0 @@
-package com.hanghe.redis.mysql.seat
-
-import com.hanghe.redis.movie.seat.SeatEntity
-
-interface SeatRepository {
-
-    fun findByTheaterIdAndSeatIdsIn(theaterId: Long, seatIds: List<Long>): List<SeatEntity>
-
-    fun findByScreeningIdAndSeatIds(screeningId: Long, seatIds: List<Long>): List<SeatEntity>
-
-    fun findByTheaterIdAndSeatIds(theaterId: Long, seatIds: List<Long>): List<SeatEntity>
-}
