--- infrastructure/src/main/kotlin/com/hanghe/redis/mysql/theater/TheaterJpaRepository.kt
@@ -1,6 +1,0 @@
-package com.hanghe.redis.mysql.theater
-
-import com.hanghe.redis.theater.TheaterEntity
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface TheaterJpaRepository : JpaRepository<TheaterEntity, Long>
