--- infrastructure/src/main/kotlin/com/hanghe/redis/mysql/theater/TheaterRepository.kt
@@ -1,8 +1,0 @@
-package com.hanghe.redis.mysql.theater
-
-import com.hanghe.redis.theater.TheaterEntity
-
-interface TheaterRepository {
-
-    fun findAll(): List<TheaterEntity>
-}
