--- infrastructure/src/main/kotlin/com/hanghe/redis/mysql/theater/TheaterRepositoryImpl.kt
@@ -1,14 +1,0 @@
-package com.hanghe.redis.mysql.theater
-
-import com.hanghe.redis.theater.TheaterEntity
-import org.springframework.stereotype.Repository
-
-@Repository
-class TheaterRepositoryImpl(
-    private val theaterJpaRepository: TheaterJpaRepository
-) : TheaterRepository {
-
-    override fun findAll(): List<TheaterEntity> {
-        return theaterJpaRepository.findAll()
-    }
-}
