+++ settings.gradle.kts
@@ -9,3 +9,7 @@ include("common")
 include("domain")
 include("infrastructure")
 include("application")
+include("infrastructure:mysql")
+include("infrastructure:message")
+include("infrastructure:cache")
+include("infrastructure:ratelimiter")
